refactor: clean up and remove unused components from modal tabs
- Remove unused ReferenceLinesSettings import from ParametersTab
- Add scrollable container for parameter rows in YAxisGroup
- Remove focus dropdown trigger behavior in YParametersSettings
- Simplify ReferenceLineTab to use existing ReferenceLinesSettings component
