+++ components/charts/EditModal/parameters/ParametersTab.tsx
@@ -1,10 +1,9 @@
 "use client"
 
-import React, { useState } from "react"
+import React from "react"
 import { ChartComponent } from "@/types"
 import { XParameterSettings } from "./XParameterSettings"
 import { YParametersSettings } from "./YParametersSettings"
-import { ReferenceLinesSettings } from "./ReferenceLinesSettings"
 
 interface ParametersTabProps {
   editingChart: ChartComponent
@@ -12,21 +11,6 @@ interface ParametersTabProps {
 }
 
 export function ParametersTab({ editingChart, setEditingChart }: ParametersTabProps) {
-
-  interface ReferenceLineConfig {
-    id: string
-    type: "vertical" | "horizontal"
-    label: string
-    xValue?: string
-    yValue?: string
-    yRangeMin?: string
-    yRangeMax?: string
-    xRangeMin?: string
-    xRangeMax?: string
-  }
-
-  const [referenceLines, setReferenceLines] = useState<ReferenceLineConfig[]>([])
-
   return (
     <div className="flex flex-col space-y-4 h-full">
       <XParameterSettings 
@@ -37,12 +21,6 @@ export function ParametersTab({ editingChart, setEditingChart }: ParametersTabPr
       <YParametersSettings 
         editingChart={editingChart} 
         setEditingChart={setEditingChart} 
-      />
-      
-      <ReferenceLinesSettings 
-        editingChart={editingChart} 
-        referenceLines={referenceLines}
-        onUpdateReferenceLines={setReferenceLines}
       />
     </div>
   )
